Print size_t variables properly

+++ test/quic_stream_test.c
@@ -20,7 +20,7 @@ static int compare_iov(const unsigned char *ref, size_t ref_len,
         total_len += iov[i].buf_len;
 
     if (ref_len != total_len) {
-        fprintf(stderr, "# expected %lu == %lu\n", ref_len, total_len);
+        fprintf(stderr, "# expected %llu == %llu\n", (long long)ref_len, (long long)total_len);
         return 0;
     }
